LayoutTests/http/tests/websocket/tests/frame-length-longer-than-buffer-expected.txt

Make sure WebSocket correctly skip lengthed frame even if received data has incomplete frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket is open
received:hello
WebSocket is closed
1
PASS areArraysEqual(received_messages, expected_messages) is true
PASS successfullyParsed is true

TEST COMPLETE
